requirements.txt
Django==2.0.0
django-ckeditor==6.0.0
django-mdeditor==0.1.18
Pillow==7.2.0
PyMySQL==0.10.0
xadmin==2.0.1
